hammer actually global now
