+++ Assets/Prefabs/GlobalManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 16acd0a4056a11b4da3e5330f300e83c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
