+++ Assets/Scripts/GlobalManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ef8d03cb85749594fa76aef92d20d1c3
