+++ Assets/Scripts/Hammer/HammerBehaviour.cs
@@ -18,38 +18,17 @@ namespace Hammer
     public class HammerBehaviour : MonoBehaviour
     {
 
-        public static HammerBehaviour Instance { get; private set; }
-
+        Quaternion imuData;
         SerialPort stream;
+        
         [SerializeField] private bool closePort;
 
         internal bool open;
         [SerializeField] private int one = 1, two = 3, three = 2, four = 4;
 
-
-        [SerializeField] private Quaternion attitude;
-
         private readonly int timeoutMs = 50;
-        [SerializeField] private float x = 0, y = 0, z = 0;
-        private static Vector3 eulerRotationAdjustment;
 
-        [SerializeField] private Quaternion CalibrationQuaternion = new Quaternion(0, 0, 0, 0);
 
-        public void CalibrateHammer()
-        {
-            CalibrationQuaternion = transform.rotation;
-        }
-        private void Awake()
-        {
-            if (Instance != null && Instance != this)
-            {
-                Destroy(gameObject);
-                return;
-            }
-
-            Instance = this;
-            DontDestroyOnLoad(gameObject);
-        }
         // Start is called once before the first execution of Update after the MonoBehaviour is created
         void Start()
         {
@@ -80,11 +59,14 @@ namespace Hammer
             }
         }
 
+        public void CalibrateHammer()
+        {
+            GlobalManager.Instance.CalibrationQuaternion = imuData;
+        }
 
         // Update is called once per frame
         void Update()
         {
-            eulerRotationAdjustment = new Vector3(x, y, z);
             if (closePort)
             {
                 open = false;
@@ -111,11 +93,11 @@ namespace Hammer
                     return;
                 }
 
-                Quaternion imuData = new Quaternion(float.Parse(quaternionString[one]), 
+                imuData = new Quaternion(float.Parse(quaternionString[one]), 
                                                     float.Parse(quaternionString[two]), 
                                                     float.Parse(quaternionString[three]), 
                                                     float.Parse(quaternionString[four]));
-                transform.rotation = imuData * CalibrationQuaternion;
+                transform.rotation = Quaternion.Inverse(GlobalManager.Instance.CalibrationQuaternion) * imuData;
                 
 
 
@@ -132,55 +114,17 @@ namespace Hammer
             }
         }
 
-        //Assert.IsTrue(WiimoteManager.HasWiimote(), "A Wiimote must be connected");
-
-        ////TODO As well as making this more efficient, we can probs use the "slow mode" booleans to improve accuracy
-        //int ret;
-        //Vector3 gyroOffset = Vector3.zero;
-        //Vector3 accelOffset = Vector3.zero;
-
-        //do {
-        //    ret = WiimoteGlobal.wiimote.ReadWiimoteData();
-
-        //    //ACCELEROMETER
-
-        //    Vector3 accelDataForFrameTest = new Vector3(
-        //        WiimoteGlobal.wiimote.Accel.GetCalibratedAccelData()[0],
-        //        WiimoteGlobal.wiimote.Accel.GetCalibratedAccelData()[1],
-        //        WiimoteGlobal.wiimote.Accel.GetCalibratedAccelData()[2]);
-        //    print("Accel data: "+accelDataForFrameTest);
-        //    accelOffset += accelDataForFrameTest;
-
-        //    //this is a test basically
-
-        //    //GYROSCOPE
-        //    //add all detected rotations throughout the frame to gyroOffset
-        //    //we should integrate this! would give accurate total rotation
-        //    if (ret > 0 && WiimoteGlobal.wiimote.current_ext == ExtensionController.MOTIONPLUS) {
-        //        gyroOffset += new Vector3(  -WiimoteGlobal.wiimote.MotionPlus.PitchSpeed,
-        //                                        -WiimoteGlobal.wiimote.MotionPlus.RollSpeed,
-        //                                        -WiimoteGlobal.wiimote.MotionPlus.YawSpeed);
-        //    }
-        //} while (ret > 0);
-
-        //gyroOffset /= 95f;
-
-        //transform.Rotate(gyroOffset, Space.Self);
-
-
-
 
         public void OnCollisionEnter(Collision collision)
         {
 
             if (collision.gameObject.CompareTag("Enemy"))
             {
-
                 Destroy(collision.gameObject);
             }
         }
 
-        void OnApplicationQuit()
+        void OnDisable ()
         {
             open = false;
             stream.Close();
